refactor(template): split template types into modules
